Redo React as a Context/Provider
The first version was basically just a useEffect hook providing a point
where we could inject the script tag wrapped in a component.

Ideally we not only want to use the upstream JS widget under the hood
but also provide other hooks to users to create more complex logic. A
provider allows us to implement this via Provider values.